--- pages/admin/delete/task.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Task = () => {
-  return (
-    <div>Task</div>
-  )
-}
-
-export default Task
